refactor: 컴포넌트 분리
refactor: 스텝 1~4 컴포넌트 분리
feat: Footer 컴포넌트 수정 및 정보 추가
style: UI 개선
