+++ lib/types/Genome.ts
