+++ package.json
@@ -9,9 +9,11 @@
     "lint": "eslint"
   },
   "dependencies": {
+    "lucide-react": "^0.562.0",
     "next": "16.1.1",
     "react": "19.2.3",
     "react-dom": "19.2.3",
+    "react-textarea-autosize": "^8.5.9",
     "zustand": "^5.0.9"
   },
   "devDependencies": {
